Widen raylib getTime() to Float64: it is a clock source
Caught by test 442 once implicit numeric conversions became an error:
uiNowSeconds() returns Float64 but its raylib branch returned getTime(),
which was Float. The compiler now refuses that silent widening, which is
exactly the point — it forced the question of what getTime() should be.

It is a clock, so Float64. Seconds-since-init in f32 degrades to ~8ms steps
after a few hours of uptime, which is visible frame-timing error in a
long-running app. The C impl and prototype return double to match.

+++ compiler/runtime/rae_runtime.h
@@ -705,7 +705,7 @@ void rae_ext_setWindowSize(int64_t width, int64_t height);
 void rae_ext_setWindowPosition(int64_t x, int64_t y);
 int64_t rae_ext_getWindowPositionX(void);
 int64_t rae_ext_getWindowPositionY(void);
-float rae_ext_getTime(void);
+double rae_ext_getTime(void);
 void rae_ext_drawCircle(float x, float y, float radius, Color color);
 void rae_ext_drawCircleGradient(int64_t x, int64_t y, float radius, Color color1, Color color2);
 void rae_ext_drawRectangle(float x, float y, float width, float height, Color color);
